requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dependency-injector==4.41.0
fastapi==0.112.2
h11==0.14.0
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.0
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
